+++ examples/playground/bin/scratch.ts
@@ -1,6 +1,6 @@
 import 'dotenv/config'
 
-import { ZoomInfoClient } from '@agentic/stdlib'
+import * as stdlib from '@agentic/stdlib'
 import restoreCursor from 'restore-cursor'
 
 /**
@@ -135,15 +135,26 @@ async function main() {
   //   category: 'linkedin profile'
   // })
 
-  const zoomInfo = new ZoomInfoClient()
-  const res = await zoomInfo.enrichContact({
-    // emailAddress: 'travis@transitivebullsh.it'
-    fullName: 'Kevin Raheja',
-    companyName: 'HeyGen'
-  })
+  // const zoomInfo = new ZoomInfoClient()
+  // const res = await zoomInfo.enrichContact({
+  //   // emailAddress: 'travis@transitivebullsh.it'
+  //   fullName: 'Kevin Raheja',
+  //   companyName: 'HeyGen'
+  // })
   // const res = await zoomInfo.searchContacts({
   //   fullName: 'Kevin Raheja'
   // })
+  // const gravatar = new stdlib.GravatarClient()
+  // const res = await gravatar.getProfileByIdentifier('email@example.com')
+
+  const hn = new stdlib.HackerNewsClient()
+  // const res = await hn.getSearchItem('43201417')
+  const res = await hn.searchItems({
+    query: 'OpenAI',
+    tags: ['story'],
+    sortBy: 'recency',
+    numericFilters: ['points>1000']
+  })
 
   console.log(JSON.stringify(res, null, 2))
 }