+++ packages/core/src/utils.ts
@@ -189,7 +189,8 @@ export function throttleKy(
 
 /**
  * Creates a new `URLSearchParams` object with all values coerced to strings
- * that correctly handles arrays of values as repeated keys.
+ * that correctly handles arrays of values as repeated keys (or CSV) and
+ * correctly removes `undefined` keys and values.
  */
 export function sanitizeSearchParams(
   searchParams:
@@ -198,7 +199,12 @@ export function sanitizeSearchParams(
         string | number | boolean | string[] | number[] | boolean[] | undefined
       >
     | object,
-  { csv = false }: { csv?: boolean } = {}
+  {
+    csv = false
+  }: {
+    /** Whether or not to use comma-separated-values for arrays or multiple entries (the default) */
+    csv?: boolean
+  } = {}
 ): URLSearchParams {
   const entries = Object.entries(searchParams).flatMap(([key, value]) => {
     if (key === undefined || value === undefined) {