+++ packages/jina/src/jina-client.ts
@@ -116,10 +116,12 @@ export class JinaClient extends AIFunctionsProvider {
 
   constructor({
     apiKey = getEnv('JINA_API_KEY'),
+    timeoutMs = 60_000,
     throttle = true,
     ky = defaultKy
   }: {
     apiKey?: string
+    timeoutMs?: number
     throttle?: boolean
     ky?: KyInstance
   } = {}) {
@@ -131,12 +133,14 @@ export class JinaClient extends AIFunctionsProvider {
       ky = ky.extend({ headers: { Authorization: `Bearer ${apiKey}` } })
     }
 
+    ky = ky.extend({ timeout: timeoutMs })
+
     const throttledKyReader = throttle
       ? throttleKy(
           ky,
           pThrottle({
             limit: apiKey ? 200 : 20,
-            interval: 60 * 60 * 1000
+            interval: 60 * 60 * 1000 // 60 minutes
           })
         )
       : ky
@@ -147,13 +151,16 @@ export class JinaClient extends AIFunctionsProvider {
           ky,
           pThrottle({
             limit: apiKey ? 40 : 5,
-            interval: 60 * 60 * 1000
+            interval: 60 * 60 * 1000 // 60 minutes
           })
         )
       : ky
-    this.kySearch = throttledKySearch.extend({ prefixUrl: 'https://s.jina.ai	' })
+    this.kySearch = throttledKySearch.extend({ prefixUrl: 'https://s.jina.ai' })
   }
 
+  /**
+   * Reads the contents of the given URL and returns it's main contents in a clean, LLM-friendly format.
+   */
   @aiFunction({
     name: 'readUrl',
     description:
@@ -194,6 +201,9 @@ export class JinaClient extends AIFunctionsProvider {
     }
   }
 
+  /**
+   * Searches the web for the given query and returns the top-5 results including their page contents in a clean, LLM-friendly format.
+   */
   @aiFunction({
     name: 'search',
     description: