+++ packages/jina/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @agentic/jina
 
+## 7.3.5
+
+### Patch Changes
+
+- Add gravatar; minor type fixes
+
+## 7.3.4
+
+### Patch Changes
+
+- Fix increase default JinaClient timeoutMs to 60s
+
 ## 7.3.3
 
 ### Patch Changes