+++ packages/stdlib/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @agentic/stdlib
 
+## 7.4.5
+
+### Patch Changes
+
+- Fix gravatar types
+- Updated dependencies
+  - @agentic/gravatar@7.4.3
+
+## 7.4.4
+
+### Patch Changes
+
+- Add gravatar; minor type fixes
+- Updated dependencies
+  - @agentic/gravatar@7.4.2
+  - @agentic/jina@7.3.5
+  - @agentic/social-data@7.3.4
+
+## 7.4.3
+
+### Patch Changes
+
+- Fix increase default JinaClient timeoutMs to 60s
+- Updated dependencies
+  - @agentic/jina@7.3.4
+
+## 7.4.2
+
+### Patch Changes
+
+- Bug fix for zoominfo re-authentication
+- Updated dependencies
+  - @agentic/zoominfo@7.4.2
+
 ## 7.4.1
 
 ### Patch Changes